[Mobile] [Android] Rename back the appId
