+++ mobile/android/app/_BUCK
@@ -35,12 +35,12 @@ android_library(
 
 android_build_config(
     name = "build_config",
-    package = "com.righton.android",
+    package = "com.rightonnew",
 )
 
 android_resource(
     name = "res",
-    package = "com.righton.android",
+    package = "com.rightonnew",
     res = "src/main/res",
 )
