+++ mobile/android/app/src/main/java/com/rightonnew/components/MainComponentsRegistry.java
@@ -1,4 +1,4 @@
-package com.righton.android.components;
+package com.rightonnew.components;
 
 import com.facebook.jni.HybridData;
 import com.facebook.proguard.annotations.DoNotStrip;
@@ -19,10 +19,8 @@ public class MainComponentsRegistry {
     SoLoader.loadLibrary("fabricjni");
   }
 
-  @DoNotStrip private final HybridData mHybridData;
-
   @DoNotStrip
-  private native HybridData initHybrid(ComponentFactory componentFactory);
+  private final HybridData mHybridData;
 
   @DoNotStrip
   private MainComponentsRegistry(ComponentFactory componentFactory) {
@@ -33,4 +31,7 @@ public class MainComponentsRegistry {
   public static MainComponentsRegistry register(ComponentFactory componentFactory) {
     return new MainComponentsRegistry(componentFactory);
   }
+
+  @DoNotStrip
+  private native HybridData initHybrid(ComponentFactory componentFactory);
 }
